+++ beacon-impl/src/main/java/fr/woorib/beacon/persistance/Store.java
@@ -1,5 +1,6 @@
 package fr.woorib.beacon.persistance;
 
+import java.math.BigDecimal;
 import java.util.List;
 import fr.woorib.beacon.data.BeaconEntry;
 
@@ -7,7 +8,7 @@ import fr.woorib.beacon.data.BeaconEntry;
  * Created by Veryeld on 30/03/2017.
  */
 public interface Store {
-    Integer saveBeacon(Integer userId, Double latitude, Double longitude);
+    Integer saveBeacon(Integer userId, BigDecimal latitude, BigDecimal longitude);
 
     BeaconEntry getBeaconByBeaconId(Integer beaconId);
